Fix response parsing for RawCall
